Add a status block to all CRDs created by AbstractOperator
This initial PR creates an empty status block in all CRDs
created by the AbstractOperator, regardless of json schema.
Additionally, it adds a simple routine to optionallyupdate the
status block to contain a 'state' string and 'lastTransitionTime'
but is API compatible with earlier versions of AstractOperator.

In the future, more general support for definable status blocks
can be added.